{{cookiecutter.app_name}}/requirements.txt
anyio==3.4.0; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.4.1; python_full_version >= "3.6.1" and python_version >= "3.6"
certifi==2021.10.8; python_full_version >= "3.6.1"
charset-normalizer==2.0.9; python_full_version >= "3.6.1" and python_version >= "3"
click==8.0.3; python_full_version >= "3.6.1" and python_version >= "3.6"
colorama==0.4.4
dnspython==2.1.0; python_full_version >= "3.6.1" and python_version >= "3.6"
email-validator==1.1.3; python_full_version >= "3.6.1"
fastapi==0.70.1; python_full_version >= "3.6.1"
greenlet==1.1.2; python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
h11==0.12.0; python_full_version >= "3.6.1" and python_version >= "3.6"
httptools==0.2.0
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.6"
itsdangerous==2.0.1; python_full_version >= "3.6.1" and python_version >= "3.6"
jinja2==3.0.3; python_version >= "3.6"
markupsafe==2.0.1; python_full_version >= "3.6.1" and python_version >= "3.6"
orjson==3.6.5; python_full_version >= "3.6.1" and python_version >= "3.7"
pydantic==1.8.2; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
python-dotenv==0.19.2
python-multipart==0.0.5; python_full_version >= "3.6.1"
pyyaml==5.4.1
requests==2.26.0; python_full_version >= "3.6.1"
six==1.16.0; python_full_version >= "3.6.1"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
sqlalchemy2-stubs==0.0.2a19; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlalchemy==1.4.29; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
sqlmodel==0.0.5; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
starlette==0.16.0; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
ujson==4.3.0; python_full_version >= "3.6.1" and python_version >= "3.6"
urllib3==1.26.7; python_full_version >= "3.6.1" and python_version < "4"
uvicorn==0.15.0; python_full_version >= "3.6.1"
uvloop==0.16.0
watchgod==0.7
websockets==10.1
